requirements.txt
absl-py==1.4.0
altair==4.2.0
altgraph==0.17.2
astor==0.8.1
asttokens==2.2.1
astunparse==1.6.3
attrs==23.1.0
backcall==0.2.0
base58==2.1.1
blinker==1.4
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==2.1.1
click==7.1.2
colorama==0.4.6
comm==0.1.3
contourpy==1.1.0
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
entrypoints==0.4
et-xmlfile==1.1.0
executing==1.2.0
fastdtw==0.3.4
flatbuffers==23.5.26
fonttools==4.41.0
future==0.18.3
gast==0.4.0
gitdb==4.0.9
GitPython==3.1.27
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.56.0
h5py==3.9.0
idna==3.4
ipykernel==6.24.0
ipython==8.14.0
ipywidgets==8.0.1
jedi==0.18.2
Jinja2==3.1.2
joblib==1.3.1
jsonschema==4.17.3
jupyter_client==8.3.0
jupyter_core==5.3.1
jupyterlab-widgets==3.0.2
keras==2.13.1
kiwisolver==1.4.4
libclang==16.0.6
llvmlite==0.40.1
Markdown==3.4.3
MarkupSafe==2.1.3
matplotlib==3.7.2
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
numba==0.57.1
numpy==1.24.3
oauthlib==3.2.2
openpyxl==3.1.2
opt-einsum==3.3.0
packaging==23.1
pandas==2.0.3
parso==0.8.3
patsy==0.5.3
pefile==2021.9.3
pickleshare==0.7.5
Pillow==10.0.0
platformdirs==3.9.1
prompt-toolkit==3.0.39
protobuf==3.19.5
psutil==5.9.5
pure-eval==0.2.2
pyarrow==12.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydeck==0.7.1
Pygments==2.15.1
pyinstaller==4.7
pyinstaller-hooks-contrib==2021.4
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.16.0
python-dateutil==2.8.2
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
pywin32==306
pywin32-ctypes==0.2.0
pyzmq==25.1.0
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.3.0
scipy==1.11.1
seaborn==0.12.2
six==1.16.0
sklearn==0.0.post5
smmap==5.0.0
stack-data==0.6.2
statsmodels==0.14.0
streamlit==1.3.0
tensorboard-data-server==0.7.1
tensorflow-estimator==2.13.0
tensorflow-io==0.31.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
threadpoolctl==3.2.0
toml==0.10.2
toolz==0.12.0
tornado==6.3.2
traitlets==5.9.0
tslearn==0.6.1
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.1
urllib3==1.26.16
validators==0.18.2
watchdog==2.3.1
wcwidth==0.2.6
Werkzeug==2.3.6
widgetsnbextension==4.0.2
wrapt==1.15.0
